pyproject.toml
[tool.poetry]
name = "chatporto"
version = "0.1.0"
description = ""
authors = ["AnaCarolSant <164232172+AnaCarolSant@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.37.1"
python-dotenv = "^1.0.1"
langchain = "^0.2.14"
pypdf2 = "^3.0.1"
faiss-cpu = "^1.8.0.post1"
openai = "^1.41.0"
langchain-community = "^0.2.12"
tiktoken = "^0.7.0"
langchain-cli = "^0.0.29"
langchain-openai = "^0.1.22"
streamlit-chat = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
